♻️ Move level.go to public package and add aliases to definitions in it
